dom/base/test/test_bug403841.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=403841
-->
<head>
  <title>Test for Bug 403841</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=403841">Mozilla Bug 403841</a>
<span id="content">abc</p>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 403841 -  Crash in nsContentUtils::CreateContextualFragment when passed a non-element node as context node  **/
  var t = document.getElementById("content").firstChild;
  var r = document.createRange();
  r.setStart(t,0);
  r.setEnd(t,3);
  // make sure this doesn't crash
  var f = r.createContextualFragment("<span>");
  ok(f.firstChild instanceof HTMLSpanElement, "created fragment ok");
</script>
</pre>
</body>
</html>
